Zero runlen (#195)
* Support for zero detection run-lenghts in whole chunks is complete

* Add a new blosc2_chunk_zeros()

* Offsets are now stored as 128-bit values

* Fix a bug in little-endian writes to frames

* Revert back to 64-bit offsets

* Do not free the chunk when appending in frames, as it is not necessary

* Add a new version for frame format, fixing endianness of counters.

* Entire chunks that are a zero run-length are encoded in offsets now

* Implement support for zero runlens in blosc_getitem

* Go back to use default codec for offsets in frame

* Support for runlen NANs

* Blosc1 backward (but not forward!) compatibility is tested now:

* Ability to generate Blosc1 headers with the envvar BLOSC_BLOSC1_COMPAT.
* Make sure that we can read chunks from Blosc1.
* Bonus: Support for optimized repeated values (new `blosc2_chunk_repeatval()`).

* Add the math lib

* Undo the previous attempt to include math lib

* Properly add math lib

* Fix *some* warnings discovered by MSVC

* Improve tests on Blosc1 compat.  Fix some Win errors.

* Fix some Win errors.

* Do not run blosc1_compat on win

* Fix more MSVC warnings

* Add Aleix's cutest.h testing framework

* Do not attempt to run tests that require more than 4 GB on 32-bit platforms

* Reduce memory footprint of the test suite

* Continue the quest towards warning-free on MSVC

* Do a free of the chunk when appending in a frame

* Fix some leaks

* Increase the range for chunk offsets

* Documented the special values for the offsets in a frame

* Better comments

* A more complete check of special values

* Add previously removed example

* Update chunk format docs.

* Don't be unnecessaily explicit

* Small improvements in API docs

* Rewrite special value condition

* Use 4 threads when compressing offsets chunk

* Fix the bench for zero detection

* WIP: Support for 0-runlen in update/insert

* Support for 0-runlen in update/insert

* Add zero_runlen test

+++ bench/sum_openmp.c
@@ -74,7 +74,7 @@
 int main(void) {
   static DTYPE udata[N];
   DTYPE chunk_buf[CHUNKSIZE];
-  size_t isize = CHUNKSIZE * sizeof(DTYPE);
+  int32_t isize = CHUNKSIZE * sizeof(DTYPE);
   DTYPE sum, compressed_sum;
   int64_t nbytes, cbytes;
   blosc2_schunk* schunk;
